break ulid functions into separate files, test
