+++ cmd/opampsupervisor/supervisor/supervisor.go
@@ -6,7 +6,6 @@ package supervisor
 import (
 	"bytes"
 	"context"
-	"crypto/rand"
 	_ "embed"
 	"errors"
 	"fmt"
@@ -15,7 +14,6 @@ import (
 	"os"
 	"path/filepath"
 	"sort"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"text/template"
@@ -1099,29 +1097,4 @@ func (s *Supervisor) findRandomPort() (int, error) {
 	}
 
 	return port, nil
-}
-
-func loadULIDFromFile(file string) (ulid.ULID, error) {
-	by, err := os.ReadFile(file)
-	if err != nil {
-		return ulid.ULID{}, err
-	}
-
-	ulidAsString := string(by)
-	return ulid.Parse(strings.TrimSpace(ulidAsString))
-}
-
-func saveULIDToFile(file string, id ulid.ULID) error {
-	ulidAsString := id.String()
-	return os.WriteFile(file, []byte(ulidAsString), 0600)
-}
-
-func generateNewULID() (ulid.ULID, error) {
-	entropy := ulid.Monotonic(rand.Reader, 0)
-	id, err := ulid.New(ulid.Timestamp(time.Now()), entropy)
-	if err != nil {
-		return ulid.ULID{}, err
-	}
-
-	return id, nil
 }
